Update README.md with documentation skeleton
The assignments for this class will be deliberately vague, with a lot of room for expression.  This is appropriate for a senior-level CS course.  Given that the assignments themselves will vary greatly, the most important part of this (and every project) is your documentation.  Along with your project, please include documentation with [parameters on Canvas]

***

Documentation can be in any format, but as you'll be creating a github repo, a markdown readme file is probably the most universal format.  

Documentation will be graded more heavily than the actual code.  Honestly your planning, evaluation, and description of the project tells me more than the code can.  Of course, I want to see the code too.

Please be prepared to demonstrate your project briefly in class.